+++ lib/rubocop_linter_app.rb
@@ -3,10 +3,10 @@
 require 'net/http'
 require 'json'
 require 'time'
-require_relative "./report_adapter"
-require_relative "./github_check_run_service"
-require_relative "./github_client"
-require_relative "./github_app"
+require_relative './report_adapter'
+require_relative './github/check_run_service'
+require_relative './github/http_client'
+require_relative './github/app'
 
 module RubocopLinterApp
   def self.run(github_data: nil, report_path: nil)
